test(e2e): extract proxy route setup
