+++ test/e2e/live/hermes-inference-switch-helpers.ts
@@ -10,6 +10,7 @@ import path from "node:path";
 import { resolveAgentInferenceApi } from "../../../src/lib/inference/config.ts";
 import { buildAvailabilityProbeEnv } from "../fixtures/availability-env.ts";
 import type { HostCliClient } from "../fixtures/clients/host.ts";
+import { resultText } from "../fixtures/clients/index.ts";
 import {
   type SandboxClient,
   trustedSandboxShellScript,
@@ -17,7 +18,10 @@ import {
 } from "../fixtures/clients/sandbox.ts";
 import { expect } from "../fixtures/e2e-test.ts";
 import type { FakeOpenAiCompatibleServer } from "../fixtures/fake-openai-compatible.ts";
-import { DEFAULT_HOSTED_INFERENCE_MODEL } from "../fixtures/hosted-inference.ts";
+import {
+  DEFAULT_HOSTED_INFERENCE_BASE_URL,
+  DEFAULT_HOSTED_INFERENCE_MODEL,
+} from "../fixtures/hosted-inference.ts";
 import {
   closeServer,
   writeJsonResponse as jsonResponse,
@@ -51,6 +55,12 @@ export const RUNTIME_SWITCH_API =
   resolveAgentInferenceApi("hermes", SWITCH_PROVIDER, SWITCH_API) ?? SWITCH_API;
 const SWITCH_MOCK_PORT = Number.parseInt(process.env.NEMOCLAW_SWITCH_MOCK_PORT ?? "0", 10);
 const INSTALL_ATTEMPTS = process.env.CI === "true" || process.env.GITHUB_ACTIONS === "true" ? 3 : 1;
+export const PROXY_RESOLUTION_PROVIDER = "nvidia-prod";
+export const PROXY_RESOLUTION_MODEL = "nvidia/nemotron-proxy-resolution-e2e";
+export const PROXY_FORBIDDEN_MARKERS = [
+  "openshell:resolve:env:",
+  "sk-OPENSHELL-PROXY-REWRITE",
+] as const;
 
 interface MockCompatibleAnthropicProvider {
   endpointUrl: string;
@@ -156,6 +166,101 @@ export function env(apiKey?: string, extra: NodeJS.ProcessEnv = {}): NodeJS.Proc
       NEMOCLAW_PROVIDER: "custom",
     });
   return { ...out, ...extra };
+}
+
+export async function expectOpenAiProvider(
+  host: HostCliClient,
+  providerName: string,
+  credentialEnv: string,
+): Promise<void> {
+  const provider = await host.command(
+    "openshell",
+    ["provider", "get", "-g", "nemoclaw", providerName],
+    {
+      artifactName: `${providerName}-openai-provider-metadata`,
+      env: env(),
+      timeoutMs: 30_000,
+    },
+  );
+  const output = resultText(provider);
+  expect(provider.exitCode, output).toBe(0);
+  const plain = stripAnsi(output);
+  expect(plain).toMatch(/^\s*Type:\s*openai\s*$/imu);
+  expect(plain).toContain(credentialEnv);
+  expect(plain).toContain("OPENAI_BASE_URL");
+}
+
+export async function prepareProxyResolutionRoute({
+  apiKey,
+  host,
+  mockBaseline,
+  publicProvider,
+  redactionValues,
+}: {
+  apiKey: string;
+  host: HostCliClient;
+  mockBaseline: FakeOpenAiCompatibleServer | undefined;
+  publicProvider: ShellProbeResult | null;
+  redactionValues: string[];
+}): Promise<{ model: string; requestOffset: number }> {
+  const endpoint =
+    mockBaseline?.baseUrl ?? process.env.NEMOCLAW_ENDPOINT_URL ?? DEFAULT_HOSTED_INFERENCE_BASE_URL;
+  const model = mockBaseline ? PROXY_RESOLUTION_MODEL : SWITCH_MODEL;
+  const requestOffset = mockBaseline?.requests().length ?? 0;
+
+  // Hosted mode already registered and attached the exact nvidia-prod
+  // provider. The mock path needs an OpenAI provider for its local fixture.
+  if (publicProvider !== null) {
+    return { model, requestOffset };
+  }
+
+  const registered = await host.command(
+    "openshell",
+    [
+      "provider",
+      "create",
+      "-g",
+      "nemoclaw",
+      "--name",
+      PROXY_RESOLUTION_PROVIDER,
+      "--type",
+      "openai",
+      "--credential",
+      "NVIDIA_INFERENCE_API_KEY",
+      "--config",
+      `OPENAI_BASE_URL=${endpoint}`,
+    ],
+    {
+      artifactName: "register-proxy-resolution-provider",
+      env: env(undefined, { NVIDIA_INFERENCE_API_KEY: apiKey }),
+      redactionValues,
+      timeoutMs: 120_000,
+    },
+  );
+  expect(registered.exitCode, resultText(registered)).toBe(0);
+  await expectOpenAiProvider(host, PROXY_RESOLUTION_PROVIDER, "NVIDIA_INFERENCE_API_KEY");
+
+  const setRoute = await host.command(
+    "node",
+    [
+      CLI,
+      "inference",
+      "set",
+      "--provider",
+      PROXY_RESOLUTION_PROVIDER,
+      "--model",
+      model,
+      "--no-verify",
+    ],
+    {
+      artifactName: "set-proxy-resolution-route",
+      env: env(),
+      redactionValues,
+      timeoutMs: 180_000,
+    },
+  );
+  expect(setRoute.exitCode, resultText(setRoute)).toBe(0);
+  return { model, requestOffset };
 }
 
 export async function preCleanBestEffort(run: () => Promise<unknown>): Promise<void> {
